Changes to data model, method to get number of comments on articles in batch

+++ comments.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python2.7
 import sys, os
 import datetime
+import hashlib
 
 # Make this run with Apache and mod_wsgi
 if __name__ != "__main__":
@@ -12,9 +13,9 @@ if __name__ != "__main__":
 sys.path.append(os.getcwd())
 
 
-from bottle import Bottle, HTTPResponse, route, run, template, request, response, default_app
+from bottle import Bottle, HTTPResponse, request, response
 from bottle.ext import sqlalchemy
-from sqlalchemy import create_engine, Column, Integer, Sequence, String, DateTime, Text, desc
+from sqlalchemy import create_engine, Column, Integer, Sequence, String, DateTime, Text, desc, func
 from sqlalchemy.ext.declarative import declarative_base
 import json
 import bleach
@@ -23,12 +24,8 @@ from wheezy.captcha.image import captcha
 
 from wheezy.captcha.image import background
 from wheezy.captcha.image import curve
-from wheezy.captcha.image import noise
-from wheezy.captcha.image import smooth
 from wheezy.captcha.image import text
 
-from wheezy.captcha.image import offset
-from wheezy.captcha.image import rotate
 from wheezy.captcha.image import warp
 
 import config
@@ -61,12 +58,12 @@ class Comment(Base):
 
 class Captcha(Base):
     __tablename__ = 'captcha'
-    id = Column(Integer, Sequence('id_seq'), primary_key = True)
+    id = Column(String(8), primary_key = True)
     value = Column(String(8))
     date_created = Column(DateTime(), default=datetime.datetime.now)
 
     def __repr__(self):
-      return "<Captcha(id=%d, value=%s)>" % (self.id, self.value)
+      return "<Captcha(id=%s, value=%s)>" % (self.id, self.value)
 
 @app.post('/add')
 def add_comment(db):
@@ -75,20 +72,30 @@ def add_comment(db):
     article_id = request.forms.get('id')
     captcha_value = request.forms.get('captcha')
     captcha_id = request.forms.get('captcha_id')
-    
+
+    if not username:
+        return {"success": False, "missing": "username"}
+
+    if not text:
+        return {"success": False, "missing": "text"}
+
+    if not article_id or not captcha_id or not captcha_value:
+        return {"success": False, "missing": "captcha"}
+
     captcha = db.query(Captcha).get(captcha_id)
     captcha_ok = captcha_value.lower().strip() == captcha.value.lower()
-    
-    os.remove("captcha/captcha_%d.jpg" % (captcha.id))
+
+    os.remove("captcha/captcha_%s.jpg" % (captcha.id))
     db.delete(captcha)
 
     if not captcha_ok:
-        return json.dumps({"success": False})
-    
+        return {"success": False, "missing": "captcha"}
+
     c = Comment(username=bleach.clean(username), text=bleach.clean(text), article_id=article_id)
     db.add(c)
 
-    return json.dumps({"success": True})
+    return {"success": True}
+
 
 @app.route('/get/<hash_id>')
 def get_comments(hash_id, db):
@@ -102,6 +109,19 @@ def get_comments(hash_id, db):
 
     response.headers['Content-Type'] = 'application/json'
     return(json.dumps({ 'comments': comments }, default=dt_converter))
+
+
+@app.post('/count_batch/')
+def count_comments_batch(db):
+    # only accept XHR reques
+    if not config.DEBUG and not request.is_xhr:
+        return
+
+    ids = json.load(request.body)
+    comment_count = db.query(Comment.article_id, func.count(Comment.article_id)).filter(Comment.article_id.in_(ids)).group_by(Comment.article_id).all()
+
+    return {x: y for x, y in comment_count}
+
 
 @app.route('/count/<hash_id>')
 def count_comments(hash_id, db):
@@ -123,12 +143,13 @@ def create_captcha(db):
     old_captchas = db.query(Captcha).filter(Captcha.date_created <= dt_thresh).all()
 
     for c in old_captchas:
-        os.remove("captcha/captcha_%d.jpg" % (c.id))
+        os.remove("captcha/captcha_%s.jpg" % (c.id))
 
     db.query(Captcha).filter(Captcha.date_created <= dt_thresh).delete()
 
-    value = random.sample(string.uppercase + string.digits, 5)
-    c = Captcha(value=''.join(value))
+    value = ''.join(random.sample(string.uppercase + string.digits, 5))
+    id = hashlib.sha1(value +  datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")).hexdigest()[:8]
+    c = Captcha(value=value, id=id)
     db.add(c)
     db.flush()
 
@@ -142,16 +163,12 @@ def create_captcha(db):
                          color="#50504f",
                          drawings=[
                             warp(),
-                            #rotate(),
-                            #offset()
                     ]),
                     curve(color='#40403f'),
-                    #noise(),
-                    #smooth()
                     ])
 
     finished_img = img(value)
-    finished_img.save('captcha/captcha_%d.jpg' % (c.id), 'JPEG', quality=75)
+    finished_img.save('captcha/captcha_%s.jpg' % (c.id), 'JPEG', quality=75)
 
     return json.dumps({ "id": c.id })
 
@@ -163,7 +180,7 @@ def get_captcha(id, db):
             'Content-Type': 'image/jpeg'
         }
 
-        with open("captcha/captcha_%d.jpg" % (captcha.id)) as stream:
+        with open("captcha/captcha_%s.jpg" % (captcha.id)) as stream:
             body = stream.read()
 
         return HTTPResponse(body, **headers)
